# server/api/Service/DTOs/RequestDto/CreateAuthorRequestDto.cs
using System.ComponentModel.DataAnnotations;

namespace api.Service.DTOs.RequestDto;

public record CreateAuthorRequestDto
{
    [Required, MinLength(3)]
    public string Name { get; set; } = null!;
}