# server/api/openapi-with-docs.json
{
  "x-generator": "NSwag v14.6.0.0 (NJsonSchema v11.5.0.0 (Newtonsoft.Json v13.0.0.0))",
  "openapi": "3.0.0",
  "info": {
    "title": "My Title",
    "version": "1.0.0"
  },
  "paths": {
    "/api/Author": {
      "get": {
        "tags": [
          "Author"
        ],
        "operationId": "Author_GetAuthors",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuthorDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Author"
        ],
        "operationId": "Author_CreateAuthor",
        "requestBody": {
          "x-name": "req",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAuthorRequestDto"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthorDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Author/{id}": {
      "get": {
        "tags": [
          "Author"
        ],
        "operationId": "Author_GetAuthorById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthorDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Author"
        ],
        "operationId": "Author_UpdateAuthor",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "req",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAuthorRequestDto"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthorDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Author"
        ],
        "operationId": "Author_DeleteAuthor",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/Book": {
      "get": {
        "tags": [
          "Book"
        ],
        "operationId": "Book_GetBooks",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Book"
        ],
        "operationId": "Book_CreateBook",
        "requestBody": {
          "x-name": "req",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBookRequestDto"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Book/{id}": {
      "get": {
        "tags": [
          "Book"
        ],
        "operationId": "Book_GetBookById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Book"
        ],
        "operationId": "Book_UpdateBook",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "req",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBookRequestDto"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Book"
        ],
        "operationId": "Book_DeleteBook",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/Genre": {
      "get": {
        "tags": [
          "Genre"
        ],
        "operationId": "Genre_GetGenres",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GenreDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Genre"
        ],
        "operationId": "Genre_CreateGenre",
        "requestBody": {
          "x-name": "req",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGenreRequestDto"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenreDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Genre/{id}": {
      "get": {
        "tags": [
          "Genre"
        ],
        "operationId": "Genre_GetGenreById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenreDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Genre"
        ],
        "operationId": "Genre_UpdateGenre",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "req",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGenreRequestDto"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenreDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Genre"
        ],
        "operationId": "Genre_DeleteGenre",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AuthorDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "createdat": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "bookIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "CreateAuthorRequestDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 3
          }
        }
      },
      "UpdateAuthorRequestDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "newName"
        ],
        "properties": {
          "newName": {
            "type": "string",
            "minLength": 3
          },
          "booksIds": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "string"
            }
          }
        }
      },
      "BookDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "pages": {
            "type": "integer",
            "format": "int32"
          },
          "createdat": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "genre": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/GenreDto"
              }
            ]
          },
          "authorsIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "GenreDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "createdat": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "books": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "CreateBookRequestDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "title",
          "pages"
        ],
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1
          },
          "pages": {
            "type": "integer",
            "format": "int32",
            "maximum": 2147483647.0,
            "minimum": 1.0
          },
          "genreId": {
            "type": "string",
            "nullable": true
          },
          "authorsIds": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "string"
            }
          }
        }
      },
      "UpdateBookRequestDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "newTitle": {
            "type": "string",
            "minLength": 1,
            "nullable": true
          },
          "newPageCount": {
            "type": "integer",
            "format": "int32",
            "maximum": 2147483647.0,
            "minimum": 1.0,
            "nullable": true
          },
          "genreId": {
            "type": "string",
            "nullable": true
          },
          "authorsIds": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "string"
            }
          }
        }
      },
      "CreateGenreRequestDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 3
          }
        }
      },
      "UpdateGenreRequestDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "newName"
        ],
        "properties": {
          "newName": {
            "type": "string",
            "minLength": 3
          }
        }
      }
    }
  }
}